cilium: set packet_type to ensure packet is not dropped by stack
The dmac may not be set in the case of !HAVE_FIB_SUPPORT so we need
to ensure the stack doesn't mark this packet as OTHERHOST and drop
it.

Easiest way to do this in my opinion is to simply force the packet
type PACKET_HOST. Here I don't bother to guard this for encryption
or not encryption and do it in all cases.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>